rework on back for token logic
