+++ back/endpoints/db_query.py
@@ -2,7 +2,6 @@ import os
 from fastapi import APIRouter, Depends
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from back.endpoints.auth import authorize_user
 from back.utils.query import Query
 from back.api_model import QueryModel
 
@@ -14,6 +13,10 @@ DB_URL = os.getenv("DATABASE_URL", "leo:postgres@localhost:5432/sporting")
 conn = create_engine("postgresql://" + DB_URL)
 Session = sessionmaker(conn)
 session = Session()
+
+
+def authorize_user():
+    pass
 
 
 @db_router.get("/simple_query/")
