+++ back/main.py
@@ -7,7 +7,6 @@ from fastapi.responses import JSONResponse, FileResponse
 from fitparse import FitFile
 import pandas as pd
 from sqlalchemy import create_engine
-from back.endpoints.auth import router, authorize_user
 from back.endpoints.db_query import db_router
 from back.endpoints.comments import activity_router
 from back.data.etl.running_feeder import RunningFeeder
@@ -18,6 +17,7 @@ from back.data.etl.futur_wkt_feeder import FuturWorkoutFeeder
 from back.data.tables import Base
 from back.utils.data_handler import get_data
 from back.fit.fit_writer import WorkoutWriter
+from back.utils.utilities import authorize_user
 from back.api_model import (
     EventModel,
     ThresholdModel,
@@ -27,7 +27,6 @@ from back.api_model import (
 
 DB_URL = os.getenv("DATABASE_URL", "leo:postgres@localhost:5432/sporting")
 app = FastAPI()
-app.include_router(router)
 app.include_router(db_router)
 app.include_router(activity_router)
